@Watson1978 2018-03-18
Hmm, CI has performed tests on Ruby 2.3.6. 
Regexp#match? and String#match? were introduced at Ruby 2.4.0 :(