@kou 2018-03-31
How about using refinements?

```ruby
unless String.method_defined?(:match?)
  module CSV
    module MatchP
      refine String do
        def match?(pattern)
          pattern === self
        end
      end
    end

    refine Regexp do
      def match?(string)
        self === string
      end
    end
  end

  using CSV::MatchP
end

# ...
```

If it doesn't any negative result on Ruby 2.3, 2.4, 2.5 and trunk, we can merge this safely.