@Watson1978 2018-04-01
@kou I added workaround by your idea to solve Ruby 2.3 issue. Thank you.