@kou 2018-04-01
Can you show benchmark result on Ruby 2.3 and trunk?
If this change sill faster on trunk and not changed on Ruby 2.3, I'll merge this.