@Watson1978 2018-04-02
OK, I took a benchmark result.
Seems there is no performance regression for Ruby 2.3.
(The original data is in https://gist.github.com/Watson1978/323367e52db63b8ab8ec46152a4b8390)

## Ruby 2.5.1
  | Before | After | Speed up
-- | -- | -- | --
CSV.foreach | 44.641 | 47.635 | 6.7%
CSV#shift | 45.630 | 48.476 | 6.2%
CSV.read | 38.527 | 44.661 | 15.9%
CSV.table | 10.293 | 11.247 | 9.3%

## Ruby 2.4.4
  | Before | After | Speed up
-- | -- | -- | --
CSV.foreach | 42.116 | 45.793 | 8.7%
CSV#shift | 43.474 | 46.901 | 7.9%
CSV.read | 37.656 | 40.355 | 7.2%
CSV.table | 10.295 | 10.739 | 4.3%

## Ruby 2.3.7
  | Before | After | Speed up
-- | -- | -- | --
CSV.foreach | 40.263 | 40.450 | 0.5%
CSV#shift | 40.088 | 40.929 | 2.1%
CSV.read | 39.133 | 39.399 | 0.7%
CSV.table | 9.766 | 9.960 | 2.0%